postgresSQL/requirements.txt
Flask
flask-cors
python-dotenv
Flask-SQLAlchemy==2.5.1